Revert "Merge remote-tracking branch 'origin/feature/newrelic_instrumentation2' into feature/NR-1001"
This reverts commit d8ff42991d8b50900c1180d8d408307918b573e7.
