+++ vscode/package.json
@@ -628,7 +628,6 @@
 		"date-fns": "1.30.1",
 		"https-proxy-agent": "3.0.1",
 		"lodash": "4.17.21",
-		"node-abort-controller": "3.0.1",
 		"node-fetch": "2.6.7",
 		"uuid": "3.3.2",
 		"vscode-languageclient": "5.2.1"
