+++ vscode/src/extension.ts
@@ -1,12 +1,5 @@
 "use strict";
-import { Agent as HttpsAgent } from "https";
-import * as url from "url";
-import HttpsProxyAgent from "https-proxy-agent";
-// import * as vscode from "vscode";
-import fetch, { RequestInit } from "node-fetch";
 import { ProtocolHandler } from "protocolHandler";
-import { ScmTreeDataProvider } from "views/scmTreeDataProvider";
-import { AbortController } from "node-abort-controller";
 import {
 	Disposable,
 	env,
@@ -19,18 +12,19 @@ import {
 	window,
 	workspace
 } from "vscode";
-import { WebviewLike } from "webviews/webviewLike";
+import { ScmTreeDataProvider } from "views/scmTreeDataProvider";
 import { CodeStreamWebviewSidebar } from "webviews/webviewSidebar";
+import { WebviewLike } from "webviews/webviewLike";
 
 import { CodemarkType } from "@codestream/protocols/api";
-
-import { GitExtension } from "./@types/git";
+// import * as vscode from "vscode";
 import {
 	CreatePullRequestActionContext,
 	GitLensApi,
 	HoverCommandsActionContext,
 	OpenPullRequestActionContext
 } from "./@types/gitlens";
+import { GitExtension } from "./@types/git";
 import { SessionStatus, SessionStatusChangedEvent } from "./api/session";
 import { ContextKeys, GlobalState, setContext } from "./common";
 import { Config, configuration, Configuration } from "./configuration";
@@ -89,31 +83,6 @@ export async function activate(context: ExtensionContext) {
 
 		cfg = configuration.get<Config>();
 	}
-	let telemetryOptions;
-	try {
-		const proxyAgent = getHttpsProxyAgent(getInitializationOptions());
-		const requestInit: RequestInit | undefined = {
-			agent: proxyAgent,
-			headers: {
-				"X-CS-Plugin-IDE": IDE_NAME
-			}
-		};
-
-		// 5s failsafe
-		const controller = new AbortController();
-		const timeout = setTimeout(() => {
-			controller.abort();
-		}, 5000);
-		requestInit.signal = controller.signal;
-
-		const response = await fetch(`${cfg.serverUrl.trim()}/no-auth/nr-ingest-key`, requestInit);
-
-		clearTimeout(timeout);
-
-		telemetryOptions = await response.json();
-	} catch (ex) {
-		Logger.warn(`no NewRelic telemetry - ${ex.message}`);
-	}
 
 	let webviewLikeSidebar: (WebviewLike & CodeStreamWebviewSidebar) | undefined = undefined;
 	// this plumping lives here rather than the WebviewController as it needs to get activated here
@@ -153,8 +122,7 @@ export async function activate(context: ExtensionContext) {
 			traceLevel: Logger.level,
 			machineId: env.machineId
 		},
-		webviewLikeSidebar,
-		telemetryOptions
+		webviewLikeSidebar
 	);
 
 	const scmTreeDataProvider = new ScmTreeDataProvider();
@@ -494,78 +462,4 @@ async function showStartupUpgradeMessage(version: string, previousVersion: strin
 			// }
 		}
 	}
-}
-
-export function getHttpsProxyAgent(options: {
-	proxySupport?: string;
-	proxy?: {
-		url: string;
-		strictSSL?: boolean;
-	};
-}) {
-	let _httpsAgent: HttpsAgent | HttpsProxyAgent | undefined = undefined;
-	const redactProxyPasswdRegex = /(http:\/\/.*:)(.*)(@.*)/gi;
-	if (
-		options.proxySupport === "override" ||
-		(options.proxySupport == null && options.proxy != null)
-	) {
-		if (options.proxy != null) {
-			const redactedUrl = options.proxy.url.replace(redactProxyPasswdRegex, "$1*****$3");
-			Logger.log(
-				`Proxy support is in override with url=${redactedUrl}, strictSSL=${options.proxy.strictSSL}`
-			);
-			_httpsAgent = new HttpsProxyAgent({
-				...url.parse(options.proxy.url),
-				rejectUnauthorized: options.proxy.strictSSL
-			} as any);
-		} else {
-			Logger.log("Proxy support is in override, but no proxy settings were provided");
-		}
-	} else if (options.proxySupport === "on") {
-		const proxyUrl = process.env.HTTPS_PROXY || process.env.HTTP_PROXY;
-		if (proxyUrl) {
-			const strictSSL = options.proxy ? options.proxy.strictSSL : true;
-			const redactedUrl = proxyUrl.replace(redactProxyPasswdRegex, "$1*****$3");
-			Logger.log(`Proxy support is on with url=${redactedUrl}, strictSSL=${strictSSL}`);
-
-			let proxyUri;
-			try {
-				proxyUri = url.parse(proxyUrl);
-			} catch {}
-
-			if (proxyUri) {
-				_httpsAgent = new HttpsProxyAgent({
-					...proxyUri,
-					rejectUnauthorized: options.proxy ? options.proxy.strictSSL : true
-				} as any);
-			}
-		} else {
-			Logger.log("Proxy support is on, but no proxy url was found");
-		}
-	} else {
-		Logger.log("Proxy support is off");
-	}
-	return _httpsAgent;
-}
-
-export function getInitializationOptions(
-	options: { proxy?: any; proxySupport?: string; newRelicTelemetryEnabled?: boolean } = {}
-) {
-	if (Container.config.proxySupport !== "off") {
-		const httpSettings = workspace.getConfiguration("http");
-		const proxy = httpSettings.get<string | undefined>("proxy", "");
-		if (proxy) {
-			options.proxy = {
-				url: proxy,
-				strictSSL: httpSettings.get<boolean>("proxyStrictSSL", true)
-			};
-			options.proxySupport = "override";
-		} else {
-			options.proxySupport = "on";
-		}
-	} else {
-		options.proxySupport = "off";
-	}
-
-	return options;
 }
