+++ vscode/src/webviews/webviewSidebar.ts
@@ -117,39 +117,14 @@ export class CodeStreamWebviewSidebar implements WebviewLike, Disposable, Webvie
 				if (err) {
 					reject(err);
 				} else {
-					let html = data.replace(
-						/{{root}}/g,
-						Uri.file(Container.context.asAbsolutePath("."))
-							.with({ scheme: "vscode-resource" })
-							.toString()
+					resolve(
+						data.replace(
+							/{{root}}/g,
+							Uri.file(Container.context.asAbsolutePath("."))
+								.with({ scheme: "vscode-resource" })
+								.toString()
+						)
 					);
-					if (
-						Container.telemetryOptions &&
-						Container.telemetryOptions.browserIngestKey &&
-						Container.telemetryOptions.accountId &&
-						Container.telemetryOptions.applicationId
-					) {
-						try {
-							const browserScript = Container.context.asAbsolutePath("dist/newrelic-browser.js");
-							if (browserScript) {
-								const browser = fs
-									.readFileSync(browserScript)
-									.toString()
-									.replace(/{{accountID}}/g, Container.telemetryOptions.accountId!)
-									.replace(/{{applicationID}}/g, Container.telemetryOptions.applicationId!)
-									// agentID === applicationID
-									.replace(/{{agentID}}/g, Container.telemetryOptions.applicationId!)
-									.replace(/{{licenseKey}}/g, Container.telemetryOptions.browserIngestKey);
-								html = html.replace(
-									"<head>",
-									`<head><script type="text/javascript">${browser}</script>`
-								);
-							}
-						} catch (ex) {
-							Logger.log("NewRelic telemetry", { error: ex });
-						}
-					}
-					resolve(html);
 				}
 			});
 		});
